Modify D.cs , Change int to long

+++ ABC139/D.cs
@@ -7,7 +7,7 @@ namespace ABC139
     {
         static void Main(string[] args)
         {
-            int N = int.Parse(Console.ReadLine());
+            long N = int.Parse(Console.ReadLine());
             long res = 0;
             for (long i = 0; i < N; i++)
             {
